Refactor ( Rmved old files and moved code around )
